refactor: move stencil-server main package to server/ directory
