--- main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/odpf/stencil/server"
-)
-
-func main() {
-	server.Main()
-}
