+++ server/api/ping_test.go
@@ -1,17 +1,17 @@
 package api_test
 
 import (
+	server2 "github.com/odpf/stencil/server/server"
 	"net/http"
 	"net/http/httptest"
 	"testing"
 
-	"github.com/odpf/stencil/server"
 	"github.com/odpf/stencil/server/api"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestPing(t *testing.T) {
-	router := server.Router(&api.API{})
+	router := server2.Router(&api.API{})
 	w := httptest.NewRecorder()
 	req, _ := http.NewRequest("GET", "/ping", nil)
 	router.ServeHTTP(w, req)
