+++ server/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/odpf/stencil/server/server"
+)
+
+func main() {
+	server.Start()
+}
