+++ server/scripts/build.sh
